Re-introduce dd-model and introduce dd-bind
dd-model didn't work after some rewrite long ago, and I strongly thought
about removing it as it was quite hacky. And in say a login form I would
_strongly_ suggest to not bind to any model at all, simply post the form.
Also binding to models doesn't support the auto-fill by browsers.
But in web apps it can be really really really handy to bind to inputs.
It's also handy to call a trigger, while listening for another path,
which is why I introduced the much more complex and superhacky dd-bind.
dd-bind has a bounce and reset timeout as well, which should make it
possible to work well with async data.